Big cleanup. Marked beta. Moved some UI files.  Removed all need for ReflectionUtils (SongLoader IPA patch adds what we need). Removed unused imports;
