+++ SongBrowserPlugin/SongBrowserPlugin.csproj
@@ -77,9 +77,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Logger.cs" />
-    <Compile Include="ReflectionUtil.cs" />
     <Compile Include="SongBrowserApplication.cs" />
-    <Compile Include="SongBrowserMasterViewController.cs" />
+    <Compile Include="UI\SongBrowserMasterViewController.cs" />
     <Compile Include="Plugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SongBrowserModel.cs" />
