+++ SongBrowserPlugin/UIBuilder.cs
@@ -37,11 +37,6 @@ namespace SongBrowserPlugin
         {
             T vc = new GameObject(name).AddComponent<T>();
 
-            /*vc.rectTransform.anchorMin = new Vector2(0f, 0f);
-            vc.rectTransform.anchorMax = new Vector2(1f, 1f);
-            vc.rectTransform.sizeDelta = new Vector2(0f, 0f);
-            vc.rectTransform.anchoredPosition = new Vector2(0f, 0f);*/
-
             return vc;
         }
